Moved the AV API calls
